% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_check_data.R
\name{api_check_data}
\alias{api_check_data}
\title{Create or update MS Bioscreen data?}
\usage{
api_check_data(sufl_data, ignore_col_names = c("first_name", "last_name"),
  endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE,
  overwrite_na_to_missing = FALSE)
}
\arguments{
\item{sufl_data}{a data frame of 1 row with appropriate identifier columns; see 'Details'.}

\item{ignore_col_names}{the column names to ignore when comparing \code{sufl_data} with the data in the bioscreen. Set to NULL to compare all columns.}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}

\item{overwrite_na_to_missing}{if sufl_data contains an NA, should this data be ignored or should this value be considered
missing? The default is set to FALSE so that any NA values will be ignored; see 'Details'.}
}
\value{
\code{api_check_data} returns 'create' if there is no data in the bioscreen that matches the identifiers found in sufl_data. If there is data in
the bioscreen that matches the identifiers found in sufl_data, \code{api_check_data} returns either 'update' or 'no action'. \code{api_check_data}
returns 'no action' when all sufl_data values match those in the bioscreen entry. Note that if there are missing values in non-identifier columns
of the bioscreen that are non-missing in sufl_data, \code{api_check_data} will return 'update'. If there are NA values in non-identifier columns
of sufl_data that are non-missing in the bioscreen entry, \code{api_check_data} will return 'update' only if \code{overwrite_na_to_missing} is TRUE.
}
\description{
\code{api_check_data} searches bioscreen for data entry that matches the identifiers of the inputed data frame and indicates
whether bioscreen data needs to be updated, created, or left alone based on the inputed data frame.
}
\details{
sufl_data must be a data frame of 1 row with column names that follow the SUFL format. For checking subjects data, sufl_data needs to
contain the columns, 'source_id' and 'external_identifier'. For checking attacks, treatments, or visits data, sufl_data needs to
contain the columns, 'source_id', 'external_identifier', 'patient_source_id', and 'patient_external_identifier'. If any of these
column names are missing, \code{api_check_data} will return an error message. Also note that if there are NA values in non-identifier
columns of sufl_data that are non-missing in the bioscreen, \code{api_check_data} will only return 'update' if
\code{overwrite_na_to_missing} is TRUE.
}
\seealso{
\code{\link{api_get}}, \code{\link{api_create}}, \code{\link{api_update}}
}

