% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_delete.R
\name{api_delete}
\alias{api_delete}
\title{Delete entry in MS Bioscreen}
\usage{
api_delete(source_id = 1, external_identifier = 100,
  endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE)
}
\arguments{
\item{source_id}{source identifer}

\item{external_identifier}{external identifer}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_delete} deletes specified entry in MS Bioscreen.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}}
}

