% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_delete_batch.R
\name{api_delete_batch}
\alias{api_delete_batch}
\title{Delete MS Bioscreen data entries}
\usage{
api_delete_batch(source_ids, endpoint,
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{source_ids}{source identifers}

\item{endpoint}{the data endpoint of interest. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_delete_batch} deletes all the bioscreen data entries associated with the specified \code{source_ids},
and \code{endpoint}.
}
\details{
Deleting 'subjects' data will also delete the associated 'attacks', 'treatments', and 'visits' data.
Deleting 'attacks', 'treatments', or 'visits' data will NOT delete the associated data in any of the other
endpoints.
}
\seealso{
\code{\link{api_delete}}
}

