% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_do_action.R
\name{api_do_action}
\alias{api_do_action}
\title{General API interaction tool}
\usage{
api_do_action(action = GET,
  url = "https://msbioscreen-uat.herokuapp.com/api/v1/subjects",
  token = get_token(), json_body_data = NULL, results_per_page = NULL)
}
\arguments{
\item{action}{the HTTP method (GET, DELETE, PUT, POST).}

\item{url}{the url of the page to retrieve.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{json_body_data}{a non-array JSON. This JSON may be used for searching, updating, or creating data. If NULL (default),
the body argument is not part of the HTTP request header.}

\item{results_per_page}{number of results to return per page. If set to NULL (default), the results_per_page query argument
in the HTTP request header is equal to its default value.}
}
\description{
\code{api_do_action} is a general API interaction tool that can fetch, delete, update, create, and search
bioscreen data.
}
\seealso{
\code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}},
\code{\link{api_search_by_epicid}}, \code{\link{api_search_by_updated_at}}
\code{\link{api_check}}
}

