% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_get.R
\name{api_get}
\alias{api_get}
\title{Get a specific entry from the MS Bioscreen.}
\usage{
api_get(source_id = 1, external_identifier = 100, endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{source_id}{source identifer}

\item{external_identifier}{external identifer}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_get} returns the bioscreen data associated with the specified \code{source_id},
\code{external_identifier}, and \code{endpoint}.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_search_by_epicid}}, \code{\link{api_search_by_updated_at}},
\code{\link{api_check}}
}

