% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_get_batch.R
\name{api_get_batch}
\alias{api_get_batch}
\title{Get batch from MS Bioscreen}
\usage{
api_get_batch(endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE,
  number_of_pages = NULL, results_per_page = NULL)
}
\arguments{
\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}

\item{number_of_pages}{number of pages of data to return. If set to NULL (default), all pages are returned.}

\item{results_per_page}{number of results to return per page. If set to NULL (default), the results_per_page query argument
in the HTTP request header is equal to its default value.}
}
\description{
\code{api_get_batch} fetches a specific data set from the MS Bioscreen.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{return_status}}, \code{\link{response_to_data_frame}}
}

