% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_search.R
\name{api_search}
\alias{api_search}
\title{General API search function}
\usage{
api_search(search_by_list = list(source_id = 1, external_identifier = 1),
  endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{search_by_list}{list of search parameter names and values}

\item{endpoint}{the data endpoint of interest. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_search} returns the bioscreen entries that match the specified search query.
}
\seealso{
\code{\link{api_get}}, \code{\link{api_search_by_msid}}, \code{\link{api_search_by_updated_at}}
}

