% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_search_by_epicid.R
\name{api_search_by_epicid}
\alias{api_search_by_epicid}
\title{Search MS Bioscreen data for specified EPICID}
\usage{
api_search_by_epicid(epicid = 1, endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE)
}
\arguments{
\item{epicid}{EPIC Identifier}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_search_by_epicid} searches bioscreen for specified \code{epicid} and returns all entries associated with that
\code{epicid}
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}}, \code{\link{api_search_by_updated_at}}
}

