% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_search_by_updated_at.R
\name{api_search_by_updated_at}
\alias{api_search_by_updated_at}
\title{Search the MS Bioscreen by when it was updated}
\usage{
api_search_by_updated_at(updated_at_date = "2015-09-24",
  updated_at_time = "00:00:00", endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token("msbwaiter_token"), verbose_b = TRUE)
}
\arguments{
\item{updated_at_date}{date stamp in format "yyy-mm-dd".}

\item{updated_at_time}{time stamp in format "hh:mm:ss".}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to search global environment and environment for the variable 'msbwaiter_token'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_search_by_updated_at} returns the bioscreen entries updated on or after the specified date and time.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}}, \code{\link{api_search_by_epicid}}
}

