% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_update.R
\name{api_update}
\alias{api_update}
\title{Update MS Bioscreen data entry}
\usage{
api_update(sufl_data, endpoint, ignore_colnames = c("first_name",
  "last_name"), base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE, overwrite_na_to_missing = FALSE)
}
\arguments{
\item{sufl_data}{a data frame with 1 row and column names that follow the current SUFL specification (1.0). At minimum, sufl_data
must contain the identifier columns "source_id" and "external_identifier". For attacks, treatments and visits data, sufl_data
must also contain the identifier columns "patient_source_id" and "patient_external_identifier".}

\item{endpoint}{the data endpoint of interest. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{ignore_colnames}{the names of the variables whose values should not be updated in the bioscreen. Set to NULL to allow any
variable to be updated in the bioscreen if its value is different in the inputed sufl data set.}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_update} updates an entry in the bioscreen.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_delete}}
}

