% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_tokens.R
\name{get_token}
\alias{get_token}
\title{Search for authorization token or prompt for it.}
\usage{
get_token(overwrite = FALSE)
}
\arguments{
\item{overwrite}{A logical. Should we prompt if token exists? Usual use
by other functions calls for the default behaviour \code{overwrite = FALSE}.}
}
\description{
\code{get_token} searches for the environment variable MSBWAITER_TOKEN. If not
set, it prompts the user for an authorization token. If the session is not
interactive, the function will throw an error.
}
\seealso{
\code{\link[base]{Sys.setenv}}
}

