% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_tokens.R
\name{prompt_for_token}
\alias{prompt_for_token}
\title{Prompt for authorization token}
\usage{
prompt_for_token()
}
\description{
\code{prompt_for_token} prompts the user for an authorization token. The token
entered is set as an evironment variable named 'msbwaiter_token'. If the user has
set 'msbwaiter_token' in the global environment, \code{prompt_for_token} will return an error.
}
\seealso{
\code{\link[base]{Sys.setenv}}, \code{\link[base]{get_token}}
}

