% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_low_level_helpers.R
\name{response_to_data_frame}
\alias{response_to_data_frame}
\title{Convert an HTTP response to a data frame}
\usage{
response_to_data_frame(response_data)
}
\arguments{
\item{response_data}{HTTP response.}
}
\description{
\code{response_to_data_frame} converts an HTTP response to a data frame
}
\seealso{
\code{\link[jsonlite]{fromJSON}}, \code{\link{api_get_batch}},
\code{\link{api_get}}, \code{\link{api_search_by_epicid}},
\code{\link{api_search_by_updated_at}}
}

