% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_low_level_helpers.R
\name{return_status}
\alias{return_status}
\title{HTTP request status}
\usage{
return_status(response_data, ok_status = c(200, 201, 202),
  print_when_ok = "Done.\\n")
}
\arguments{
\item{response_data}{HTTP response.}

\item{ok_status}{the status code that indicates the HTTP request was successful.}

\item{print_when_ok}{the message to print if the HTTP request was successful.}
}
\value{
For successful HTTP requests, \code{return_status} returns TRUE and prints the \code{print_when_ok} message. For
unsuccessful HTTP requests, \code{return_status} returns FALSE and prints the content of the response message.
}
\description{
\code{return_status} checks whether an HTTP request was successful and prints the contents of the error if
the request was not successful.
}
\seealso{
\code{\link{api_get_batch}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}}, \code{\link{api_search_by_epicid}},
\code{\link{api_search_by_updated_at}}, \code{\link{api_check_data}}
}

