% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_low_level_helpers.R
\name{to_json_non_array}
\alias{to_json_non_array}
\title{Convert R objects to JSON non-array}
\usage{
to_json_non_array(x, overwrite_na_to_missing = FALSE, ...)
}
\arguments{
\item{x}{the object to be encoded}

\item{overwrite_na_to_missing}{if x contains an NA, should this data be removed from the JSON object
or does the NA represent a missing value that should be represented as null in the JSON. The default is set to
FALSE so that any NA values will be ignored and not part of the JSON.}

\item{...}{arguments passed on to class specific print methods}
}
\value{
a JSON without brackets.
}
\description{
\code{to_json_non_array} converts an R object to a JSON non-array.
}
\seealso{
\code{\link[jsonlite]{toJSON}}, \code{\link{api_get}},
\code{\link{api_search_by_epicid}}, \code{\link{api_search_by_updated_at}}
}

