% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_check.R
\name{api_check}
\alias{api_check}
\title{Create or update MS Bioscreen data entry?}
\usage{
api_check(sufl_data, ignore_colnames = c("first_name", "last_name"), endpoint,
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE, overwrite_na_to_missing = FALSE,
  change = FALSE)
}
\arguments{
\item{sufl_data}{a data frame of 1 row with column names that follow the current SUFL specification (1.0); see 'Details'.}

\item{ignore_colnames}{the columns to remove from sufl data before proceeding. Set to NULL to keep all columns.}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}

\item{overwrite_na_to_missing}{if the sufl data set contains an NA, should this data be ignored or should
this value be encoded as a missing value? The default is set to FALSE so that any NA values will be
ignored; see 'Details'.}

\item{change}{should the function proceed to actually update or create data in the bioscreen?
Default is set to FALSE.}
}
\value{
\code{api_check} returns 'create' if there is currently no data in the bioscreen that matches the
identifier columns of sufl_data. If there is data in the bioscreen that matches the identifier columns of
sufl_data, \code{api_check} will return either 'no action' or 'update'. \code{api_check} returns
'no action' when all sufl_data values (minus the ignore_colnames) match those in the corresponding bioscreen
entry. Note that if there are missing values in the non-identifier/non-ignore_colnames columns of the bioscreen
that are non-missing in sufl_data, \code{api_check} will return 'update'. However, if there are NA values in the
non-identifier/non-ignore_colnames columns of sufl_data that are non-missing in the bioscreen,
\code{api_check} will return 'update' only if \code{overwrite_na_to_missing} is TRUE.
}
\description{
\code{api_check} searches the bioscreen for a data entry that matches the identifiers
of the inputed data frame. If no such data entry exists, \code{api_check} will return 'create'.
Otherwise, \code{api_check} will return 'update' or 'no action' depending on how closely the
values of the inputed data frame match those in the corresponding bioscreen data entry. To check many
data entries at once, use \code{api_check_batch}.
}
\details{
In order for the bioscreen API to process sufl_data, sufl_data must contain, at minimum, the SUFL identifier columns 'source_id'
and 'external_identifier'. For attacks, treatments, and visits data, sufl_data must also contain the SUFL identifier columns 'patient_source_id'
and 'patient_external_identifier'. Note that if there are NA values in the non-identifier/non-ignore_colnames columns of sufl_data that are non-missing in the
bioscreen, \code{api_check} will return 'update' only if \code{overwrite_na_to_missing} is TRUE.
}
\seealso{
\code{\link{api_check_batch}}, \code{\link{api_get}}, \code{\link{api_create}},
\code{\link{api_update}}, \code{\link{to_json_non_array}}
}

