% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_create.R
\name{api_create}
\alias{api_create}
\title{Create new entry in MS Bioscreen}
\usage{
api_create(sufl_data, endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE)
}
\arguments{
\item{sufl_data}{a data frame of 1 row with appropriate identifier columns; see 'Details'.}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_create} creates a new entry in MS Bioscreen.
}
\details{
sufl_data must be a data frame of 1 row with column names that follow the SUFL format. For creating subjects data, sufl_data needs to
contain the columns, 'source_id' and 'external_identifier'. For creating attacks, treatments, or visits data, sufl_data needs to
contain the columns, 'source_id', 'external_identifier', 'patient_source_id', and 'patient_external_identifier'. If any of these
column names are missing, \code{api_create} will return an error message.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_update}}
}

