% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_create.R
\name{api_create}
\alias{api_create}
\title{Create a new entry in the MS Bioscreen}
\usage{
api_create(sufl_data, endpoint,
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{sufl_data}{a data frame of 1 row with column names that follow the current SUFL specification (1.0); see 'Details'.}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_create} creates a new entry in the bioscreen.
}
\details{
sufl_data must be a data frame of 1 row with column names that follow the SUFL format. For creating
subjects data, sufl_data needs to contain the identifier columns, 'source_id' and 'external_identifier'.
For creating attacks, treatments, or visits data, sufl_data needs to contain the identifier columns,
'source_id', 'external_identifier', patient_source_id', and 'patient_external_identifier'. If any of
these columns are missing, \code{api_create} will return an error message.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_update}}, \code{\link{api_delete}}
}

