% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_delete.R
\name{api_delete}
\alias{api_delete}
\title{Delete MS Bioscreen data entry}
\usage{
api_delete(source_id = 1, external_identifier = 100, endpoint,
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{source_id}{source identifer}

\item{external_identifier}{external identifer}

\item{endpoint}{the data endpoint of interest. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_delete} deletes the bioscreen data associated with the specified \code{source_id},
\code{external_identifier}, and \code{endpoint}.
}
\details{
Deleting 'subjects' data will also delete the associated 'attacks', 'treatments', and 'visits' data.
Deleting 'attacks', 'treatments', or 'visits' data will NOT delete the associated data in any of the other
endpoints.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}}
}

