% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_search_by_epicid.R
\name{api_search_by_epicid}
\alias{api_search_by_epicid}
\title{Search MS Bioscreen for a specific EPICID}
\usage{
api_search_by_epicid(epicid = 1, endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = get_token(), verbose_b = TRUE)
}
\arguments{
\item{epicid}{EPIC identifier}

\item{endpoint}{the data endpoint of interest. Possible values are "subjects", "attacks",
"treatments", and "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is to get environment variable 'MSBWAITER_TOKEN'.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_search_by_epicid} returns the bioscreen entries associated with the specified \code{epicid}.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}}, \code{\link{api_search_by_updated_at}}
}

