% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_search_by_updated_at.R
\name{api_search_by_updated_at}
\alias{api_search_by_updated_at}
\title{Search MS Bioscreen data by when it was updated}
\usage{
api_search_by_updated_at(updated_at = "2015-09-24 00:00:00",
  endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE)
}
\arguments{
\item{updated_at}{time stamp in format "yyy-mm-dd hh:mm:ss".}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}
}
\description{
\code{api_search_by_epicid} searches bioscreen for all entries updated on or after specified \code{updated_at} and returns these entries.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}, \code{\link{api_update}},
\code{\link{api_delete}}, \code{\link{api_get}}, \code{\link{api_search_by_epicid}}
}

