% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_update.R
\name{api_update}
\alias{api_update}
\title{Update entry in MS Bioscreen}
\usage{
api_update(sufl_data, endpoint = "subjects",
  base_url = "https://msbioscreen-uat.herokuapp.com/api/v1",
  token = Sys.getenv("msbwaiter_token"), verbose_b = TRUE,
  overwrite_na_to_missing = FALSE)
}
\arguments{
\item{sufl_data}{a data frame of 1 row with appropriate identifier columns; see 'Details'.}

\item{endpoint}{the name of the data endpoint to fetch. Possible values are "subjects", "attacks", "treatments", "visits".}

\item{base_url}{the API base URL.}

\item{token}{HTTP authorization token. Default is the token that the user enters when loading msbwaiter.}

\item{verbose_b}{print progress messages as function runs?}

\item{overwrite_na_to_missing}{if sufl_data contains an NA, should this data be ignored or should this value be considered
missing? The default is set to FALSE so that any NA values will be ignored; see 'Details'.}
}
\description{
\code{api_update} updates a specified entry in the MS Bioscreen.
}
\details{
sufl_data must be a data frame of 1 row with column names that follow the SUFL format. For updating subjects data, sufl_data needs to
contain the columns, 'source_id' and 'external_identifier'. For updating attacks, treatments, or visits data, sufl_data needs to
contain the columns, 'source_id', 'external_identifier', 'patient_source_id', and 'patient_external_identifier'. If any of these
column names are missing, \code{api_update} will return an error message. Also note that if there are NA values in non-identifier
columns of sufl_data that are non-missing in the bioscreen, \code{api_update} will only update bioscreen data to missing if
\code{overwrite_na_to_missing} is TRUE.
}
\seealso{
\code{\link{api_do_action}}, \code{\link{api_create}}
}

