% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_tokens.R
\name{get_token}
\alias{get_token}
\title{Search for authorization token}
\usage{
get_token(token_name = "msbwaiter_token")
}
\description{
\code{get_token} searches the global environment for \code{token_name} and returns its value if it exists. If
\code{token_name} has not been set in the global environment, \code{get_token} will search the environment
for \code{token_name} and return its value if it exisits. If \code{token_name} is not found in either environment,
\code{get_token} will return an error.
}
\seealso{
\code{\link[base]{Sys.setenv}}, \code{\link[base]{prompt_for_token}}
}

